requirements.txt
# Career Radar — Python dependencies
# Install: pip install -r requirements.txt

feedparser>=6.0.11
anthropic>=0.40.0
python-dotenv>=1.0.1
streamlit>=1.39.0
pandas>=2.2.0
